refactor: detach dependencies on Navigation
