+++ src/components/molecules/Navigation.tsx
@@ -1,24 +1,14 @@
-import { useEffect, useRef } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { useQuery } from 'react-query';
-import { Button, Flex, Text } from 'rebass';
+import { Flex } from 'rebass';
 
-import usePostEventLog from '~hooks/api/usePostEventLog';
-import useScrollTo from '~hooks/useScrollTo';
-import { sendAmplitudeData } from '~utils/amplitude';
+interface NavigationProps {
+  children: React.ReactNode;
+}
 
-const Navigation = () => {
-  const ref = useRef<HTMLElement>();
-
-  const { handleExecuteScroll } = useScrollTo('test');
-
+const Navigation = ({ children }: NavigationProps) => {
   const { data: isMobile } = useQuery<boolean>('isMobile');
-
-  const { mutateEventLog } = usePostEventLog();
-
-  const { data } = useQuery<{ testType: string }>('source');
-
-  const { data: utmSource } = useQuery('utmSource');
-
+  const ref = useRef<HTMLElement>();
   const handleScroll = () => {
     let lastScrollTop = 0;
     const checkScrolLDirection = () => {
@@ -37,13 +27,6 @@ const Navigation = () => {
     };
     checkScrolLDirection();
   };
-
-  const handleClick = () => {
-    sendAmplitudeData('버튼클릭_검사하기__네비게이션', { source: data?.testType, utmSource });
-    handleExecuteScroll();
-    mutateEventLog('CLICK_CTA_BUTTON');
-  };
-
   useEffect(() => {
     window.addEventListener('scroll', handleScroll);
 
@@ -51,15 +34,9 @@ const Navigation = () => {
       window.removeEventListener('scroll', handleScroll);
     };
   }, []);
-
   return (
     <Flex ref={ref} as="nav" variant="navigation">
-      <Text fontWeight={800} fontSize="30px" flex={2} color="primary">
-        DEVTI
-      </Text>
-      <Button variant="primary" fontWeight={700} onClick={handleClick}>
-        검사하기
-      </Button>
+      {children}
     </Flex>
   );
 };
